Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
werkzeug==2.3.7
pymongo==4.5.0
pymongo[srv]==4.5.0